// frontend/src/api/models/PracticeQuestion.ts
export interface PracticeQuestion {
    /**
     * 
     * @type {number}
     * @memberof PracticeQuestion
     */
    position: number;
    /**
     * 
     * @type {QuestionType}
     * @memberof PracticeQuestion
     */
    questionType: QuestionType;
    /**
     * 
     * @type {string}
     * @memberof PracticeQuestion
     */
    question: string;
    /**
     * 
     * @type {string}
     * @memberof PracticeQuestion
     */
    correctAnswer?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PracticeQuestion
     */
    exampleAnswer?: string | null;
    /**
     * 
     * @type {Array<PracticeOption>}
     * @memberof PracticeQuestion
     */
    closedOptions?: Array<PracticeOption>;
    /**
     * 
     * @type {Array<QuestionKeyword>}
     * @memberof PracticeQuestion
     */
    openKeywords?: Array<QuestionKeyword>;
    /**
     * 
     * @type {number}
     * @memberof PracticeQuestion
     */
    questionId: number;
    /**
     * 
     * @type {number}
     * @memberof PracticeQuestion
     */
    practiceId: number;
}